reorganize match files
