+++ src/02-tagger/04-correction/matches/07-verbs.js
@@ -1,0 +1,89 @@
+const verbs = '(pat|wade|ollie|will|rob|buck|bob|mark|jack)'
+
+let list = [
+  //he quickly foo
+  { match: '#Noun #Adverb [#Noun]', group: 0, tag: 'Verb', reason: 'quickly-foo' },
+  //will secure our
+  { match: 'will [#Adjective]', group: 0, tag: 'Verb', reason: 'will-adj' },
+  //he left
+  { match: '#Noun #Adverb? [left]', group: 0, tag: 'PastTense', reason: 'left-verb' },
+  //he disguised the thing
+  { match: '#Pronoun [#Adjective] #Determiner #Adjective? #Noun', group: 0, tag: 'Verb', reason: 'he-adj-the' },
+  //march to
+  { match: '[march] (up|down|back|to|toward)', group: 0, tag: 'Infinitive', reason: 'march-to' },
+  //must march
+  { match: '#Modal [march]', group: 0, tag: 'Infinitive', reason: 'must-march' },
+  //was walking
+  { match: `[#Copula (#Adverb|not)+?] (#Gerund|#PastTense)`, group: 0, tag: 'Auxiliary', reason: 'copula-walking' },
+  //support a splattering of auxillaries before a verb
+  { match: `[(has|had) (#Adverb|not)+?] #PastTense`, group: 0, tag: 'Auxiliary', reason: 'had-walked' },
+  //would walk
+  { match: `[(#Modal|did) (#Adverb|not)+?] #Verb`, group: 0, tag: 'Auxiliary', reason: 'modal-verb' },
+  //would have had
+  {
+    match: `[#Modal (#Adverb|not)+? have (#Adverb|not)+? had (#Adverb|not)+?] #Verb`,
+    group: 0,
+    tag: 'Auxiliary',
+    reason: 'would-have',
+  },
+  //would be walking
+  { match: `#Modal (#Adverb|not)+? be (#Adverb|not)+? #Verb`, group: 0, tag: 'Auxiliary', reason: 'would-be' },
+  //had been walking
+  {
+    match: `(#Modal|had|has) (#Adverb|not)+? been (#Adverb|not)+? #Verb`,
+    group: 0,
+    tag: 'Auxiliary',
+    reason: 'had-been',
+  },
+
+  //jack seems guarded
+  { match: '#Singular (seems|appears) #Adverb? [#PastTense$]', group: 0, tag: 'Adjective', reason: 'seems-filled' },
+  //fall over
+  { match: '#PhrasalVerb [#PhrasalVerb]', group: 0, tag: 'Particle', reason: 'phrasal-particle' },
+  //'the can'
+  { match: '#Determiner [(can|will|may)]', group: 0, tag: 'Singular', reason: 'the can' },
+  //is mark hughes
+  { match: '#Copula [#Infinitive] #Noun', group: 0, tag: 'Noun', reason: 'is-pres-noun' },
+  //
+  { match: '[#Infinitive] #Copula', group: 0, tag: 'Noun', reason: 'inf-copula' },
+  //sometimes not-adverbs
+  { match: '#Copula [(just|alone)]$', group: 0, tag: 'Adjective', reason: 'not-adverb' },
+  //jack is guarded
+  { match: '#Singular is #Adverb? [#PastTense$]', group: 0, tag: 'Adjective', reason: 'is-filled' },
+  //is eager to go
+  { match: '#Copula [#Adjective to] #Verb', group: 0, tag: 'Verb', reason: 'adj-to' },
+  //walking is cool
+  { match: '[#Gerund] #Adverb? not? #Copula', group: 0, tag: 'Activity', reason: 'gerund-copula' },
+  //walking should be fun
+  { match: '[#Gerund] #Modal', group: 0, tag: 'Activity', reason: 'gerund-modal' },
+  //running-a-show
+  { match: '#Gerund #Determiner [#Infinitive]', group: 0, tag: 'Noun', reason: 'running-a-show' },
+
+  //'foo-up'
+  { match: '(#Verb && @hasHyphen) up', group: 0, tag: 'PhrasalVerb', reason: 'foo-up' },
+  { match: '(#Verb && @hasHyphen) off', group: 0, tag: 'PhrasalVerb', reason: 'foo-off' },
+  { match: '(#Verb && @hasHyphen) over', group: 0, tag: 'PhrasalVerb', reason: 'foo-over' },
+  { match: '(#Verb && @hasHyphen) out', group: 0, tag: 'PhrasalVerb', reason: 'foo-out' },
+
+  // would wade
+  { match: `#Modal [${verbs}]`, group: 0, tag: 'Verb', reason: 'would-mark' },
+  { match: `#Adverb [${verbs}]`, group: 0, tag: 'Verb', reason: 'really-mark' },
+  // wade smith
+  { match: `${verbs} #Person`, tag: 'Person', reason: 'rob-smith' },
+  // wade m. Cooper
+  { match: `${verbs} #Acronym? #ProperNoun`, tag: 'Person', reason: 'rob-a-smith' },
+
+  // will walk
+  { match: '[(do|does|will|have|had)] (not|#Adverb)? #Verb', group: 0, tag: 'Auxiliary', reason: 'have-had' },
+  // damn them
+  { match: '[shit] (#Determiner|#Possessive|them)', group: 0, tag: 'Verb', reason: 'swear1-verb' },
+  { match: '[damn] (#Determiner|#Possessive|them)', group: 0, tag: 'Verb', reason: 'swear2-verb' },
+  { match: '[fuck] (#Determiner|#Possessive|them)', group: 0, tag: 'Verb', reason: 'swear3-verb' },
+
+  //will be running (not copula)
+  { match: '[will #Adverb? not? #Adverb? be] #Gerund', group: 0, tag: 'Copula', reason: 'will-be-copula' },
+  //for more complex forms, just tag 'be'
+  { match: 'will #Adverb? not? #Adverb? [be] #Adjective', group: 0, tag: 'Copula', reason: 'be-copula' },
+]
+
+module.exports = list
