--- src/02-tagger/04-correction/matches/dates.js
@@ -1,22 +1,0 @@
-module.exports = [
-  // sun the 5th
-  { match: '[sun] the #Ordinal', tag: 'WeekDay', reason: 'sun-the-5th' },
-  //sun feb 2
-  { match: '[sun] #Date', group: 0, tag: 'WeekDay', reason: 'sun-feb' },
-  //1pm next sun
-  { match: '#Date (on|this|next|last|during)? [sun]', group: 0, tag: 'WeekDay', reason: '1pm-sun' },
-
-  //sat november
-  { match: '[sat] #Date', group: 0, tag: 'WeekDay', reason: 'sat-feb' },
-  //this sat
-  { match: `(in|by|before|during|on|until|after|of|within|all) [sat]`, group: 0, tag: 'WeekDay', reason: 'sat' },
-
-  //5th of March
-  { match: '#Value of #Month', tag: 'Date', reason: 'value-of-month' },
-  //5 March
-  { match: '#Cardinal #Month', tag: 'Date', reason: 'cardinal-month' },
-  //march 5 to 7
-  { match: '#Month #Value to #Value', tag: 'Date', reason: 'value-to-value' },
-  //march the 12th
-  { match: '#Month the #Value', tag: 'Date', reason: 'month-the-value' },
-]
